+++ src/vm-repair/azext_vm_repair/scripts/win-run-driver.ps1
@@ -11,7 +11,14 @@ if ($init)
 		$logFileName = "logs-$curDate.txt"
 		Set-Content -Path "./$logFileName" -Value "[Log-Start $(Get-Date)]"
 		$logFile = (Resolve-Path "./$logFileName").Path
-		
+		if($repo_fork -eq $null)
+		{
+			$repo_fork = 'Azure/repair-script-library'
+		}
+		if($repo_branch -eq $null)
+		{
+			$repo_branch = 'master'
+		}
 		$url = "https://github.com/$repo_fork/repair-script-library/zipball/$repo_branch/"
 		(new-object net.webclient).DownloadFile($url, (Join-Path $pwd 'repair-script-library.zip'))
 		Expand-Archive -Path 'repair-script-library.zip' -DestinationPath 'repair-script-library'